Fixed issue with all frames being identical + other bugs.
When run with the parameters elev, dhav or temp the program now outputs a mp4 file showing the changes in the parameter over time. However with the parameter hvel all of the frames are identical. The program would throw an index error when run but would still output a valid mp4 file, so I used except IndexError to suppress this. I also changed the label on the colour bar depending on which parameter was used.